Removing POST from node type endpoint and slight factoring on HTTP ITs
