+++ fcrepo-http-api/src/main/java/org/fcrepo/http/api/repository/FedoraRepositoryNodeTypes.java
@@ -19,9 +19,6 @@ import static javax.ws.rs.core.MediaType.APPLICATION_XHTML_XML;
 import static javax.ws.rs.core.MediaType.APPLICATION_XML;
 import static javax.ws.rs.core.MediaType.TEXT_HTML;
 import static javax.ws.rs.core.MediaType.TEXT_PLAIN;
-import static javax.ws.rs.core.Response.status;
-import static org.apache.http.HttpStatus.SC_BAD_REQUEST;
-import static org.apache.http.HttpStatus.SC_NO_CONTENT;
 import static org.fcrepo.http.commons.domain.RDFMediaType.JSON_LD;
 import static org.fcrepo.http.commons.domain.RDFMediaType.N3;
 import static org.fcrepo.http.commons.domain.RDFMediaType.N3_ALT2;
@@ -30,24 +27,17 @@ import static org.fcrepo.http.commons.domain.RDFMediaType.RDF_XML;
 import static org.fcrepo.http.commons.domain.RDFMediaType.TURTLE;
 import static org.fcrepo.http.commons.domain.RDFMediaType.TURTLE_X;
 
-import java.io.IOException;
-import java.io.InputStream;
-
 import javax.inject.Inject;
 import javax.jcr.RepositoryException;
 import javax.jcr.Session;
-import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
-import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
-import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriInfo;
 
 import org.fcrepo.http.commons.AbstractResource;
 import org.fcrepo.http.commons.responses.HtmlTemplate;
-import org.fcrepo.kernel.exception.RepositoryRuntimeException;
 import org.fcrepo.kernel.utils.iterators.RdfStream;
 import org.springframework.context.annotation.Scope;
 
@@ -76,32 +66,5 @@ public class FedoraRepositoryNodeTypes extends AbstractResource {
     @HtmlTemplate("jcr:nodetypes")
     public RdfStream getNodeTypes(@Context final UriInfo uriInfo) throws RepositoryException {
         return nodeService.getNodeTypes(session).session(session);
-    }
-
-    /**
-     * Update the NodeTypes by POSTing a CND file/fragment
-     *
-     * @return 201
-     * @throws javax.jcr.RepositoryException
-     * @throws java.io.IOException
-     */
-    @POST
-    @Consumes({"text/cnd"})
-    @Timed
-    public Response updateCnd(final InputStream requestBodyStream)
-        throws RepositoryException, IOException {
-
-        try {
-            nodeService.registerNodeTypes(session, requestBodyStream);
-
-            return status(SC_NO_CONTENT).build();
-        } catch ( RepositoryRuntimeException ex ) {
-            // this may be brittle, but the returned exception isn't an InvalideNodeTypeDefinitionException...
-            if ( ex.getCause().getMessage().indexOf("Reading the node definitions from the"
-                    + " supplied stream resulted in problems(s)") != -1 ) {
-                return status(SC_BAD_REQUEST).entity(ex.getMessage()).build();
-            }
-            throw ex;
-        }
     }
 }
