+++ fcrepo-http-api/src/test/java/org/fcrepo/integration/http/api/UseStockConfiguration.java
@@ -1,0 +1,37 @@
+/**
+ * Copyright 2014 DuraSpace, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.fcrepo.integration.http.api;
+
+import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import org.springframework.test.context.ContextConfiguration;
+
+/**
+ * Use a stock repository configuration for tests.
+ *
+ * @author ajs6f
+ *
+ */
+@ContextConfiguration({"/spring-test/test-container.xml"})
+@Target(TYPE)
+@Retention(RUNTIME)
+public @interface UseStockConfiguration {
+
+}
