+++ fcrepo-http-api/src/test/resources/auto-versioning-web.xml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<web-app xmlns="http://java.sun.com/xml/ns/javaee"
+	 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	 xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
+	 version="3.0" metadata-complete="false">
+    
+    <display-name>Fedora-on-ModeShape</display-name>
+    
+    <context-param>
+        <param-name>contextConfigLocation</param-name>
+        <param-value>classpath:spring-test/auto-versioning-master.xml</param-value>
+    </context-param>
+
+    <listener>
+        <listener-class>org.springframework.web.context.ContextLoaderListener</listener-class>
+    </listener>
+    
+	<servlet>
+    <servlet-name>jersey-servlet</servlet-name>
+    <servlet-class>org.glassfish.jersey.servlet.ServletContainer</servlet-class>
+
+    <init-param>
+      <param-name>javax.ws.rs.Application</param-name>
+      <param-value>org.fcrepo.http.commons.FedoraApplication</param-value>
+    </init-param>
+
+    <load-on-startup>1</load-on-startup>
+	</servlet>
+ 
+	<servlet-mapping>
+		<servlet-name>jersey-servlet</servlet-name>
+		<url-pattern>/*</url-pattern>
+	</servlet-mapping>
+  
+    <!-- filter to add test auth to grizzly -->
+    <filter>
+      <filter-name>TestAuth</filter-name>
+      <filter-class>org.fcrepo.http.commons.test.util.TestAuthenticationRequestFilter</filter-class>
+    </filter>
+    
+    <filter-mapping>
+      <filter-name>TestAuth</filter-name>
+      <url-pattern>/*</url-pattern>
+    </filter-mapping>
+	
+</web-app>
