+++ fcrepo-http-api/src/test/resources/spring-test/auto-versioning-master.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="
+    http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd">
+    
+    <!-- Master context for the test application. -->
+    
+    <import resource="classpath:/spring-test/auto-versioning-repo.xml"/>
+    <import resource="classpath:/spring-test/rest.xml"/>
+    <import resource="classpath:/spring-test/eventing.xml"/>
+    <import resource="classpath:/spring-test/transactions.xml"/>
+    
+</beans>
