+++ fcrepo-http-api/src/test/resources/spring-test/auto-versioning-test-container.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xmlns:context="http://www.springframework.org/schema/context"
+  xmlns:p="http://www.springframework.org/schema/p"
+  xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
+  http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.0.xsd">
+
+  <context:property-placeholder/>
+
+  <bean id="containerWrapper"
+    class="org.fcrepo.http.commons.test.util.ContainerWrapper"
+    init-method="start" destroy-method="stop" p:port="${test.port:8080}"
+    p:configLocation="classpath:auto-versioning-web.xml"/>
+
+</beans>
